Fix corner-case in function 'construct.ranges'
When only 2 revisions are provided to the function `construct.ranges`
via parameter `revs` and `sliding.window` was `TRUE`, then `offset + 1`
exceeds the number of elements in `revs`, resulting in choosing elements
`3:2` as end elements of ranges. Instead of just constructing one range,
two ranges are constructed, which is obviously wrong. To prevent this, a
check is added whether `offset + 1` exceeds the length of `revs` and if
so, just return the last element to just construct one range.

In addition, as we are already handling corner cases, add a check at the
beginning of the function for checking whether, at least, two revisions
are specified. If there is only one revision given, no range can be
constructed, and therefore throw an error in this situation.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>