Fix cloning in data tests to use deep clones
When updating the data tests to use deep clones, the comparison of two
Conf objects brakes. This happens since a conf parameter gets a new
field 'value' when updated. When this parameter is then changed back to
the default value, this value field stays present. But since we compare
Conf objects with the comparison of their string representation, this
now brakes. The fix is to delete the 'value' field when the parameter is
set to default.

Signed-off-by: Christian Hechtl <hechtl@cs.uni-saarland.de>