Make test on eigenvector-centrality classification rebust to floating point inaccuracy
On some machines, the test using eigenvector-centrality classification failes do
to very minor differences in the accuracy of floating point numbers. Actually,
`expect_equal` is robust to such differences, but as the row order of the data
frames to compare depends on the order of the floating point centrality values,
the order of the rows could differ and, therefore, also the order of the author
names could differ due to that.

To fix this problem and make the test more robust to such inaccuracies, we add
additional checks:

(1) Check that the eigenvector centralities (i.e., one column of the data frame)
are equal (this works as `expect_equal` tolerates minor differences).
(2) Check that the sets of authornames are equal (ignoring the actual order of
the author names in the two data frames).
(3) To be able to also check whether the author names and their corresponding
centrality value match, use the following trick: Round all centrality values to
five decimal digits. Then the the values should be identical. Then reorder the
data frame columns (in order to order the author names for developers having the
same rounded centrality value consistently) and check if the resulting data are
equal.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>