Change docker containers in CI pipeline
Beginning with R version 4.1.2, the `r-base` containers on drone.io are broken,
as R cannot be executed there as of a bug in docker on debian in combination
with glibc version > 2.33.
As a workaround, we use the `rocker/r-ver` containers instead. As they have
restricted the R package repository to the latest point in time at which the R
version was most recent, we need to manually change the R repository to CRAN in
order to get the current package versions.

With this change, the due to floating point inaccuracy breaking
eigenvector-centrality test seems to work again.

´Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>