Deactivate showcase pipeline on R versions 3.4 and 3.5
Since the new release of package 'RcppEigen', which was released today,
is not anymore available for R versions 3.4 and 3.5, and since this is
only used by the `ggraph` package which is needed only when executing
the showcase file, we remove these versions from the showcase pipeline
for now until we find a workaround for this problem.
This should not have any effect on the test execution.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>