Rewrite remove.deleted.and.empty.user to take multiple columns as argument and use it in setters
The function in util-read.R now uses a vector containing the columns to
check for deleted and ampty users. This is important for commits, because
there, the column "committer.name" is also relevant.

Also use the function in the setter methods of util.data for mails,
issues, commits, and authors right after input validation.

This addresses #213.

Signed-off-by: Niklas Schneider <s8nlschn@stud.ui-saarlad.de>