Prevent R warnings in if conditions when updating configuration values
Currently, we get a couple of warnings when updating configuration values that
consist of more than one values, for example in the following case:

The default value of the configuration option of "issues.from.source" is the
following vector: `c("github", "jira")`. Then, the following statements leads to
a couple of warnings:

`project.configuration$update.value("issues.from.source", "github")`

Warnings:

'length(x) = 2 > 1' in coercion to 'logical(1)'

These warnings are caused by the default value of "issues.from.source" in
several if conditions, as
`if (is.na(private[["attributes"]][[name]][["default"]]))` returns more than one
boolean, actually two: one for "github" and one for "jira". But this is not
expected here, as the `is.na` check here should just find out whether the
overall default value is `NA` and not for all elements of a vector whether they
are `NA`. To fix this, we need to make sure that the `is.na` check is only
performed when the length of the variable to check is exactly 1.

As this check is needed multiple times, and since we also need its negated
version, I have added an additional helper function `is.single.na`, that checks
whether an element is of length 1, and if so, returns the result of `is.na`,
otherwise returns `FALSE`. This makes sure that we always have a single boolean
value in this condition, no matter whether the input is a vector of multiple
elements or not.

This prevents the warning being printed. As the the warning will be turned into
an error in future R versions, this fix also prevents errors that could be
caused by this problem.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>