Fix class of date attributes author-network test
As of igraph version 1.4.0, the classes of edge attributes are respected in
igraph. Therefore, dates in the expected outcomes of our tests need to be of
class `POSIXct`. As the `I()` operator prevents an automatic conversion, we need
to manually convert the dates to the correct class. However, as `I()` adds the
`AsIs` class, we need to remove this class from the affected edge attributes
again, as otherwise the expected network and the built network would not be
equal.

To make this test succeed in both, igraph versions < 1.4.0 and igraph versions
>= 1.4.0, the expected class is derived from the class of the date attribute in
the built network. This way, the test is able to pass independent of the used
igraph version.

This fix addresses se-sic#236.

Signed-off-by: Christian Hechtl <hechtl@cs.uni-saarland.de>
Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>