Fix the classes of the default values of vertex and edge attributes
As of igraph version 1.4.0, igraph respects the classes of edge and vertex
attributes. Therefore, we have to make sure that we add the correct classes when
we add new edge or vertex attributes to a (possibly empty) network.

To do so, in function `add.attributes.to.network`, we need to set the
`default.value` (for which we have already set the correct class before) instead
of `NA` (which is interpreted as `logical` if no concrete class is given).

In addition, if the attribute is of class `POSIXct`, an additional time zone
attribute is necessary. We use the default time zone that is set in
`util-init.R` in this case.

This fix addresses se-sic#236.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>