Fix the class of the date attribute on multi networks (due to a bug in igraph)
As of igraph version 1.4.0, igraph respects the classes of edge and vertex
attributes. However, the function `igraph::disjoint_union` does not respect
their classes and converts `POSIXct` attributes to numeric attributes, for
instance. This is a bug in igraph that is expected to be fixed in later igraph
versions. For now, we temporarily fix this problem on our own by converting the
`date` attribute of the union back to `POSIXct`.
Note: Also other attributes could be affected! However, for convenience reasons,
we only apply our temporary fix to the `date` attribute.

This addresses se-sic#236.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>