Fix edge attributes of bipartite edges
The breaking changes of igraph version 1.4.0 regarding the classes of edge
attributes revealed an actual bug in the edge attributes of bipartite edges in
coronet:

We need to make sure that we only add the edge attributes for the edges that are
really added to the network. As, for example, edges to the untracked artifact
are not added to our bipartite network, we also need to make sure to remove the
attributes of the corresponding edges (otherwise this could lead to wrong
attributes). Therefore, we add a (rather complex) check for which edges (i.e.,
vertex sequences) are present in the final bipartite network and extract the
corresponding information for the edge attributes accordingly.

This addresses se-sic#236.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>