Adjust comment regarding bug in `igraph::disjoint_union`
The bug in igraph, which we describe in a comment, affects not only igraph
versions 1.4.0 and 1.4.1, but also affects igraph version 1.4.2, as the bug
was reported after the release of version 1.4.2.
Therefore, adjust our comment accordingly.

This is a follow-up commit for commit a9535550d93207f466b315f33ea263a50e6c8924.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>