Adjust adjacency matrix to recent Matrix package versions and fix bugs
The previously used statements to initiate sparse matrices are deprecated or
disfunct in the recent versions of the Matrix package. Therefore, update the
way of how we initiate a 'dgTMatrix' (i.e., a matrix with numeric values in
triplet form). The new syntax requires, at least, Matrix version 1.3.0
(which has already been released in December 2020). Therefore, add a
corresponding warning to the install script if older versions of package
Matrix would be installed.

In addition, fix two bugs related to our expanded adjacency matrices:
- We have to call the 'which' function from the Matrix package (and not the
  'which' function of the base package).
- Replace '1:nrow' by 'seq_len' to corretly handle empty adjacency
  matrices.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>