Adjust ggsave call in showcase.R and update example plot
Since version 3.3.4 of package ggplot2, ggsave does not automatically
choose a white background any more. To stay consistent with the previously
provided example plot in the repo and keep this plot reproducible,
manually add a white background in our showcase.R.

In addition, update the example plot as there have been some slight
improvements on the legend arrows in ggplot2. The content of the example
plot has not changed, though.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>