Add empty bots list to empty test data
The empty test data contain empty files for every data source, except for the
bot data. To stay consistent with the other data sources and to also check in
the test that empty bot data are handled correctly, add empty bot data.
(This also reduces the number of warnings in the test runs.)

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>