Get directedness of issue-based artifact-networks from configuration
Instead of creating only undirected issue-based artifact-networks, we now
take the directedness information out of the network config. Edges are
already created in a way that they can be interpreted as directed
edges from the issue referencing to the referenced issue.

Replace for loop in edge creation by more efficient mclapply and fix
some minor formatting inconsistencies.

This works towards #239.

Signed-off-by: Maximilian Löffler <s8maloef@stud.uni-saarland.de>