Fix installation for Matrix package on R 3.6
As the recent package version of Matrix is not compatible any more with
R versions < 4.4.0, but the Matrix package version that is automatically
shipped with R 3.6 is smaller than the minimum package version of Matrix
that coronet requires. To circumvent this problem, specifically install
Matrix version 1.3-4 from the package archives if the minimum version
required by coronet is not met. This change should not affect recent R
versions, which automatically ship a sufficient version of Matrix.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>