Fix a bug in edgelist construction of issue networks
The bug, described in Issue#260, practially limits the amount of edges
that can be present in an issue network to 17 and additionally leads to
warnings in every case where there are not exactly 17 edges present.

It is caused by an incorrect parameter to the 'base::split' function.
We intend to pass a vector containing the numbers 1 to nrow(add.links),
but instead pass 1 to 17.

Signed-off-by: Maximilian Löffler <s8maloef@stud.uni-saarland.de>