Ensure correct data type of NA timestamps
Up until now, `get.data.cut.to.same.date(data.sources =
c("issues", "mails", "commits"))` failed if some of the first data
source was empty, but not if the second one was empty. The reason was
that `NA` values introduced by empty data sources at the beginning of
the data frame turned the data frame into a data frame of numeric
objects instead of POSIXct objects. If there were already POSIXct
objects in the data frame, this did not happen. To prevent the
timestamps to be interpreted as numeric values, make sure that the `NA`
values are always POSIXct objects.

This fixes #269.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>