Add function to compute last activity per person and activity type
While a function to compute the first activity per person and activity type
already existed, this was not the case for the last activity.

Instead of copying the code from first-activity computation, reuse the code by
establishing an additional helper function for aggregating activity dates. This
is used by first and last activity now, and could potentially also be used for
other aggregations (e.g., to compute the mid date of a person's activities).

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>