Add function to add vertex attribute "last.activity"
Similar to enhancing first-activity computation by last-activity computation in
the previous commit, this can also be done for the corresponding vertex
attributes.

We now have a new function to attribute the "last.activity" vertex attribute.
Both functions to add first and last activity as a vertex attribute make use of
a commonly used helper function that allows to add an aggregated activity date
as a vertex attribute.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>