Implement network splitting for networks with simplified edges
The attributes assigned to edges in simplified networks (potentially)
contain multiple values for a single edge. Previously, splitting such
networks was not supported as it is not obvious how to deal with such
edges.

Now, we support splitting simplified networks. An edge that comprises of
multiple source-edges before simplification should be partitioned again
and its partials distributed over the corresponding bins. Singular edges
can be handled as prior.

Signed-off-by: Maximilian Löffler <s8maloef@stud.uni-saarland.de>