Remove deprecated `scale_name` parameter from plotting module
As of `ggplot2` version 3.5.0, the function `discrete_scale`, which is used
for `plot.network` in coronet, does not require a `scale_name` parameter
anymore. Even more, this parameter has been deprecated. Therefore, we also
don't use it any more in coronet. However, as a consequence of this, the
plotting module of coronet requires `ggplot2` version 3.5.0 or higher,
as of now.

Signed-off-by: Thomas Bock <bockthom@cs.uni-saarland.de>