Remove numeric attributes from all splits and keep 'weight' in any case
As discussed in PR#278, removing numeric attributes from some splits
while keeping them for others is not benefitial. Hence, we remove them
from all splits as soon as one contains non-complete edges.

Furthermore, the 'weight' attribute is mandatory and can therefore not
be removed. Now, we assume 'weight' to always be equal to the amount of
partials that comprise an edge.

Signed-off-by: Maximilian Löffler <s8maloef@stud.uni-saarland.de>