Correctly group multi-relational and regular edges before simplification
When splifying a network without the 'simplify.multiple.relations'
options set, we must still consider that input networks may already
contain multi-relational edges (as shown in 'showcase.R' when the sample
network is simpified again).

In our recent meetings we concluded that edges with a 'mail' relation
can be simplified together with edges that have a list('mail', 'mail')
relation but not with edges that have a list('mail', 'cochange')
relation. Further, we must ensure that any multi-relational edges can
still be simplified together with multi-relational edges that have the
same relation mix.

Signed-off-by: Maximilian Löffler <s8maloef@stud.uni-saarland.de>