Ensure edge attribute values are of list type
Nomenclature: An "edge attribute" refers to the values of an attribute
over all edges of a network, while an "edge attribute value" refers to
the value of an edge attribute of a single edge.

Previously we decided that edge attributes by default are of list type,
yet we did not specify of which type edge attribute values should be.
This is especially relevant when an edge attribute value consists of
multiple values, i.e., in simplified edges, and the interplay of
simplified edges and non-simplified edges when simplifying repeatedly.

By default, simplification with the "concat" strategy wraps the values
of all source edges into a list. Therefore, we decided that not only
edge attributes should be lists, but also edge attribute values.

Signed-off-by: Maximilian Löffler <s8maloef@stud.uni-saarland.de>