Fix coverage report upload
After a change in Codecov's coverage report processing, uploaded reports
must now contain a listing of all files that could be relevant for the
report, in order to be valid.
This listing is generated by 'codecov-action' from the files in the
current working directory which implies that the current directory is
the repository. Therefore, we must checkout the repository before
triggering the 'codecov-action'.

Signed-off-by: Maximilian Löffler <s8maloef@stud.uni-saarland.de>